backend/src/main/java/ms/backend/service/MemberService.java
package ms.backend.service;

import ms.backend.domain.Member;
import ms.backend.repository.MemberRepository;
import ms.backend.repository.MemoryMemberRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;
import java.util.List;
import java.util.Optional;

//@Service
@Transactional
public class MemberService {
    private final MemberRepository memberRepository;

//    @Autowired
    public MemberService(MemberRepository memberRepository) {
        this.memberRepository = memberRepository;
    }

    /**
     * register
     */
    public Long join(Member member) {
        validateIsMemberDuplicated(member);
        memberRepository.save(member);
        return member.getId();
    }

    private void validateIsMemberDuplicated(Member member) {
        memberRepository.findByName(member.getName())
                .ifPresent(m -> {
                    throw new IllegalStateException("이미 존재하는 회원");
                });
    }

    public List<Member> findMembers() {
        return memberRepository.findAll();
    }

    public Optional<Member> findOne(Long memberId) {
        return memberRepository.findById(memberId);
    }
}
